+++ docs/changes.rst
@@ -7,6 +7,8 @@ Version 0.3.0
 To be released.
 
 - Geofront becomes to require Paramiko 1.15.0 or higher.
+- Added save check for :class:`~geofront.remote.AuthorizedKeyList`.
+  [:issue:`5`]
 
 
 Version 0.2.2